WIP: refactor client actions
