+++ src/actions.js
@@ -4,9 +4,7 @@ export {
   fetchGame,
   fetchInitialPlacement,
   fetchMoveInfo,
-  fetchMoves,
   makeMove,
-  pollMoves,
   postMove,
   selectMove,
   toggleLoading,
