+++ src/actions/pollMoves.js
@@ -1,0 +1,30 @@
+import * as client from 'services/client';
+import fetchMoves from 'actions/fetchMoves';
+
+const canUpdateMoves = ({ gameSlug, nextMovePlayer, username }) => {
+  if (gameSlug === null) return false;
+  if (username === null) return false;
+  if (username === nextMovePlayer) return false;
+
+  return true;
+};
+
+const pollMoves = ({
+  gameSlug,
+  moveCount,
+  moves,
+  nextMovePlayer,
+  username,
+}) => async dispatch => {
+  if (!canUpdateMoves({ gameSlug, nextMovePlayer, username })) return;
+
+  const {
+    data: { move_count },
+  } = await client.getMoveCount({ gameSlug });
+
+  if (moveCount >= move_count) return;
+
+  dispatch(fetchMoves({ gameSlug, moves }));
+};
+
+export default pollMoves;
